temp

+++ src/main/java/com/jdawidowska/equipmentRentalService/services/RentingService.java
@@ -14,8 +14,6 @@ import com.jdawidowska.equipmentRentalService.util.DateUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.sql.Date;
-
 @Service
 public class RentingService {
 
@@ -58,38 +56,60 @@ public class RentingService {
 
         //
         // dodajemy wpis do tabeli wypozyczen
-        //
-        RentedInventory rentedInventory = new RentedInventory();
-        rentedInventory.setIdUser(rentingRequest.getIdUser());
-        rentedInventory.setIdItem(rentingRequest.getIdItem());
-        rentedInventory.setAmount(1);
-        rentedInventory.setIdHistory(historyWithGeneratedID.getId());
-        rentedInventoryRepository.save(rentedInventory);
+        //jesli dany uzytkownik juz wypozyczyl item z takim idItem to zabdejtuj liczbe wypożyczonych itemów
+        if (rentedInventoryRepository.existsByIdUserAndIdItem(rentingRequest.getIdUser(), rentingRequest.getIdItem())) {
+            Long requiredId = rentedInventoryRepository.getIdRentedInventoryByIdUserAndIdItem(rentingRequest.getIdUser(), rentingRequest.getIdItem());
+            rentedInventoryRepository.doIncrementAmount(requiredId);
+            return true;
+        } else {
+            RentedInventory rentedInventory = new RentedInventory();
+            rentedInventory.setIdUser(rentingRequest.getIdUser());
+            rentedInventory.setIdItem(rentingRequest.getIdItem());
+            rentedInventory.setIdHistory(historyWithGeneratedID.getId());
+            rentedInventory.setAmount(1);
+            rentedInventoryRepository.save(rentedInventory);
+        }
 
         return true;
     }
 
     @Transactional
     public boolean returnItem(ReturnRequest returnRequest) {
+        //1
+        //Sprawdzamy czy dany record istnieje w RentedInventory, jesli istnieje zwracamy przedimot, jelsi nie zwracamy false
         RentedInventory rentedInventory = rentedInventoryRepository.findById(returnRequest.getIdRentedInventory()).orElse(null);
-        if(rentedInventory == null){
+        if (rentedInventory == null) {
             return false;
         }
         inventoryRepository.returnItem(rentedInventory.getIdItem());
+
+        //2
+        //dodajemy do history date oddania itemu
         UserRentHistory userRentHistory = userRentHistoryRepository.findById(rentedInventory.getIdHistory()).orElse(null);
-        if(userRentHistory == null){
+        if (userRentHistory == null) {
             return false;
         }
         userRentHistoryRepository.updateReturnDate(rentedInventory.getIdHistory(), DateUtil.getCurrentDate());
-        rentedInventoryRepository.deleteById(rentedInventory.getId());
 
-        if(returnRequest.getFeedback() != null && !returnRequest.getFeedback().isEmpty()){
+        //3
+        //usuwamy ilosc wypozyczonych rzeczy w RentedInventory
+        Integer amountRentedInventory = rentedInventory.getAmount();
+        //jesli ilosc rzeczy jest wieksza od zera to zmiejszamy ilosc o 1
+        if (amountRentedInventory == 1) {
+            //jesli ilosc jest mniejsza od zera usuwamy caly record
+            rentedInventoryRepository.deleteById(rentedInventory.getId());
+        } else {
+            rentedInventoryRepository.doDecreaseAmount(rentedInventory.getId());
+        }
+
+        //4
+        //zapisywanie feedbacku
+        if (returnRequest.getFeedback() != null && !returnRequest.getFeedback().isEmpty()) {
             Feedback feedback = new Feedback();
             feedback.setIdUser(rentedInventory.getIdUser());
             feedback.setContent(returnRequest.getFeedback());
             feedbackRepository.save(feedback);
         }
-
         return true;
     }
 }
